requirements.txt
numpy
matplotlib 
scikit-learn
scipy
tqdm
Pillow
autopep8
pillow
matplotlib
seaborn
scikit-image
sklearn
scikit-learn
holoviews[recommended]
opencv-python
torch-optimizer
torchsummary
typed-argparse
jupyterlab
joblib
tensorboard
mlflow
streamlit
pyyaml
pytorch-lightning
natsort
hydra-core==1.0.6
hydra==2.5
h5py
setuptools==59.5.0

